update: throw exceptions instead of using optional for problem and user problem service methods

+++ src/main/java/com/dsaclock/services/ProblemService.java
@@ -1,6 +1,8 @@
 package com.dsaclock.services;
 
 import com.dsaclock.entities.Problems;
+import com.dsaclock.exceptions.ProblemAlreadyExistsException;
+import com.dsaclock.exceptions.ProblemNotFoundException;
 import com.dsaclock.repos.ProblemRepo;
 import org.springframework.stereotype.Service;
 
@@ -20,22 +22,38 @@ public class ProblemService {
     }
 
     //return single problem with id
-    public Optional<Problems> getProblem(Long problemId) {
-        return problemRepo.findById(problemId);
+    public Problems getProblem(Long problemId) {
+        return problemRepo.findByProblemId(problemId).orElseThrow(() ->
+                new ProblemNotFoundException("No problem with such ID"));
     }
 
     //add new problem
     public void addProblem(Problems problem) {
+
+        if(problemRepo.findByProblemId(problem.getProblemId()).isPresent()) {
+            throw new ProblemAlreadyExistsException("Problem with this ID already exists"); //throws exception if already exists
+        }
+
         problemRepo.save(problem);
     }
 
     //update existing problem
     public void updateProblem(Problems problem) {
+
+        if(problemRepo.findByProblemId(problem.getProblemId()).isEmpty()) {
+            throw new ProblemNotFoundException("No such problem with this ID!"); //can't update non-existing problem
+        }
+
         problemRepo.save(problem);
     }
 
     //delete problem by id
     public void deleteProblem(Long problemId) {
+
+        if(problemRepo.findByProblemId(problemId).isEmpty()) {
+            throw new ProblemNotFoundException("This problem doesn't even exists bro!"); //denies deleting non-existing problem
+        }
+
         problemRepo.deleteById(problemId);
     }
 }
